+++ node_io/lib/src/internet_address.dart
@@ -9,13 +9,11 @@ import 'package:node_interop/net.dart';
 
 export 'dart:io' show InternetAddressType;
 
-const _kLoopbackIPv4 = '127.0.0.1';
-const _kLoopbackIPv6 = '::1';
-
 class InternetAddress implements io.InternetAddress {
   static const int _IPV6_ADDR_LENGTH = 16;
 
   final String _host;
+  final List<int> _inAddr;
 
   @override
   final String address;
@@ -32,9 +30,10 @@ class InternetAddress implements io.InternetAddress {
   // version does not implement this setter.
   @override
   set type(io.InternetAddressType value) =>
-      throw new UnsupportedError('Setting address type is not supported.');
+      throw new UnsupportedError('Setting address type is not allowed.');
 
-  InternetAddress._(this.address, [this._host]) {
+  InternetAddress._(this.address, [this._host])
+      : _inAddr = _inet_pton(address) {
     if (net.isIP(address) == 0)
       throw new ArgumentError('${address} is not valid.');
   }
@@ -62,31 +61,29 @@ class InternetAddress implements io.InternetAddress {
 
   @override
   bool get isLinkLocal {
-    final List<int> raw = rawAddress;
     // Copied from dart:io
     switch (type) {
       case io.InternetAddressType.IP_V4:
         // Checking for 169.254.0.0/16.
-        return raw[0] == 169 && raw[1] == 254;
+        return _inAddr[0] == 169 && _inAddr[1] == 254;
       case io.InternetAddressType.IP_V6:
         // Checking for fe80::/10.
-        return raw[0] == 0xFE && (raw[1] & 0xB0) == 0x80;
+        return _inAddr[0] == 0xFE && (_inAddr[1] & 0xB0) == 0x80;
     }
     throw new StateError('Unreachable');
   }
 
   @override
   bool get isLoopback {
-    final List<int> raw = rawAddress;
     // Copied from dart:io
     switch (type) {
       case io.InternetAddressType.IP_V4:
-        return raw[0] == 127;
+        return _inAddr[0] == 127;
       case io.InternetAddressType.IP_V6:
         for (int i = 0; i < _IPV6_ADDR_LENGTH - 1; i++) {
-          if (raw[i] != 0) return false;
+          if (_inAddr[i] != 0) return false;
         }
-        return raw[_IPV6_ADDR_LENGTH - 1] == 1;
+        return _inAddr[_IPV6_ADDR_LENGTH - 1] == 1;
     }
     throw new StateError('Unreachable');
   }
@@ -97,33 +94,85 @@ class InternetAddress implements io.InternetAddress {
     switch (type) {
       case io.InternetAddressType.IP_V4:
         // Checking for 224.0.0.0 through 239.255.255.255.
-        return rawAddress[0] >= 224 && rawAddress[0] < 240;
+        return _inAddr[0] >= 224 && _inAddr[0] < 240;
       case io.InternetAddressType.IP_V6:
         // Checking for ff00::/8.
-        return rawAddress[0] == 0xFF;
+        return _inAddr[0] == 0xFF;
     }
     throw new StateError('Unreachable');
   }
 
   @override
-  List<int> get rawAddress {
-    if (_rawAddress != null) return new List.from(_rawAddress);
-    if (type == io.InternetAddressType.IP_V4) {
-      _rawAddress = address.split('.').map(int.parse).toList(growable: false);
-    } else {
-      throw new UnimplementedError();
-    }
-    return new List.from(_rawAddress);
-  }
-
-  List<int> _rawAddress;
+  List<int> get rawAddress => new List.from(_inAddr);
 
   @override
   Future<io.InternetAddress> reverse() {
-    // dns.rever
-    throw new UnimplementedError();
+    final Completer<io.InternetAddress> completer = new Completer();
+    void reverseResult(error, List<String> hostnames) {
+      if (error != null) {
+        completer.completeError(error);
+      } else {
+        completer.complete(new InternetAddress._(address, hostnames.first));
+      }
+    }
+
+    dns.reverse(address, allowInterop(reverseResult));
+    return completer.future;
   }
 
   @override
   String toString() => '$address';
+}
+
+const int _kColon = 58;
+
+/// Best-effort implementation of native inet_pton.
+///
+/// This implementation assumes that [ip] address has been validated for
+/// correctness.
+List<int> _inet_pton(String ip) {
+  if (ip.contains(':')) {
+    // ipv6
+    final List<int> result = new List<int>.filled(16, 0);
+
+    // Special cases:
+    if (ip == '::') return result;
+    if (ip == '::1') return result..[15] = 1;
+
+    const int maxSingleColons = 7;
+
+    int totalColons = ip.codeUnits.where((code) => code == _kColon).length;
+    bool hasDoubleColon = ip.contains('::');
+    int singleColons = hasDoubleColon ? (totalColons - 1) : totalColons;
+    int skippedSegments = maxSingleColons - singleColons;
+
+    StringBuffer segment = new StringBuffer();
+    int pos = 0;
+    for (var i = 0; i < ip.length; i++) {
+      if (i > 0 && ip[i] == ':' && ip[i - 1] == ':') {
+        /// We don't need to set bytes to zeros as our [result] array is
+        /// prefilled with zeros already, so we just need to shift our position
+        /// forward.
+        pos += 2 * skippedSegments;
+      } else if (ip[i] == ':') {
+        if (segment.isEmpty) segment.write('0');
+        int value = int.parse(segment.toString(), radix: 16);
+        result[pos] = value ~/ 256;
+        result[pos + 1] = value % 256;
+        pos += 2;
+        segment.clear();
+      } else {
+        segment.write(ip[i]);
+      }
+    }
+    // Don't forget about the last segment:
+    if (segment.isEmpty) segment.write('0');
+    int value = int.parse(segment.toString(), radix: 16);
+    result[pos] = value ~/ 256;
+    result[pos + 1] = value % 256;
+    return result;
+  } else {
+    // ipv4
+    return ip.split('.').map(int.parse).toList(growable: false);
+  }
 }