+++ node_io/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0-dev.2.0
+
+- Complete implementation of `InternetAddress`.
+
 ## 1.0.0-dev.1.0
 
-- Split from node_interop
+- Split from node_interop.